+++ core/src/main/groovy/com/ainrif/apiator/core/ApiatorInfo.groovy
@@ -15,7 +15,6 @@
  */
 package com.ainrif.apiator.core
 
-import com.ainrif.apiator.core.spi.ModelTypeResolver
 import com.ainrif.apiator.core.spi.Renderer
 import com.ainrif.apiator.core.spi.WebServiceProvider
 
@@ -23,12 +22,10 @@ class ApiatorInfo {
     String version
     WebServiceProvider provider
     Renderer renderer
-    List<ModelTypeResolver> modelTypeResolvers
 
     ApiatorInfo(ApiatorConfig config) {
         this.version = ApiatorInfo.package.implementationVersion ?: 'UNPACKED'
         this.provider = config.provider
         this.renderer = config.renderer
-        this.modelTypeResolvers = config.modelTypeResolvers
     }
 }