+++ core/src/test/java/com/ainrif/apiator/core/model/api/ModelDto1.java
@@ -27,6 +27,7 @@ public class ModelDto1<TV, TVB extends Collection, GENERIC_BOUNDED extends List<
     Set<String> setField;
     String[] arrayField;
     Iterable<?> iterableField;
+    Iterable iterableFiledWOTemplateParameter;
     TV typeVariableType;
     TVB typeVariableBoundedType;
     GENERIC_BOUNDED genericBounded;