+++ renderer/core-json-renderer/plugin/spi-plugin-core-json-renderer/src/main/java/com/ainrif/apiator/renderer/plugin/spi/ModelTypePlugin.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2014-2018 Ainrif <support@ainrif.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.ainrif.apiator.renderer.plugin.spi;
+
+import javax.annotation.Nullable;
+
+public interface ModelTypePlugin extends CoreJsonRendererPlugin {
+    /**
+     * @return null if this type cannot be resolved with given resolver
+     */
+    @Nullable
+    ModelType resolve(Class<?> type);
+}