+++ src/main/groovy/gradle/GenerateJsonForUiDev.groovy
@@ -19,6 +19,7 @@ import com.ainrif.apiator.core.Apiator
 import com.ainrif.apiator.core.ApiatorConfig
 import com.ainrif.apiator.provider.jaxrs.JaxRsProvider
 import com.ainrif.apiator.renderer.core.json.CoreJsonRenderer
+import com.ainrif.apiator.renderer.plugin.jaxrs.JaxRsModelTypePlugin
 
 import javax.ws.rs.Path
 
@@ -31,7 +32,10 @@ class GenerateJsonForUiDev {
                 basePackage: jaxrsPackage,
                 apiClass: Path,
                 provider: new JaxRsProvider(),
-                renderer: new CoreJsonRenderer())
+                renderer: new CoreJsonRenderer({
+                    docletBasePackage = 'com.ainrif.apiator'
+                    plugins << new JaxRsModelTypePlugin()
+                }))
 
         new File(path).write(new Apiator(config).render())
     }